SerializerCreate serializer ancestor TrakSerializer
Fix site_details.html and site_table.html to use new address model

add create_transporter method to base class TrakSerializer

remove unneeded child to_representation methods

AddressSerializer become child of TrakSerializer

remove unused ElectronicSignature model

exclude admin.logentry and sessions.session models from runhaz.sh fixture dump

reorg, rename to be consistent
