+++ apps/api/serializers/__init__.py
@@ -1,5 +1,5 @@
 # from .addresses import AddressSerializer, MailAddressField, SiteAddressField
-from .addresses import AddressSerializer
+from .address import AddressSerializer
 from .handler import HandlerSerializer
 from .manifest import (ManifestSerializer, TransporterSerializer,
                        WasteLineSerializer)
