+++ apps/api/serializers/base/__init__.py
@@ -1,0 +1,1 @@
+from .trak import TrakSerializer
