+++ apps/api/serializers/manifest.py
@@ -3,15 +3,16 @@
 # this way if changes to the model occur, we don't need to change another
 # utility function that just reads JSON and create model instances
 
-# Comments are included in each serializer where fields have not implemented
-# this is either because ModelSerializer's defaults are sufficient or
-# they are left as a TODO item
+# Comments are included in each serializer where fields were not implemented
+# this is either because ModelSerializer's defaults are sufficient or work is needed to
+# be done
 
 from rest_framework import serializers
 
-from apps.trak.models import Address, Handler, Manifest, Transporter, WasteLine
+from apps.trak.models import Manifest, Transporter, WasteLine
 
 from . import HandlerSerializer
+from .base import TrakSerializer
 
 
 class TransporterSerializer(HandlerSerializer):
@@ -107,7 +108,7 @@ class WasteLineSerializer(serializers.ModelSerializer):
         return f'{self.lineNumber}'
 
 
-class ManifestSerializer(serializers.ModelSerializer):
+class ManifestSerializer(TrakSerializer):
     createdDate = serializers.DateTimeField(
         source='created_date')
     updatedDate = serializers.DateTimeField(
@@ -199,48 +200,23 @@ class ManifestSerializer(serializers.ModelSerializer):
         tsd_data = validated_data.pop('tsd')
         gen_data = validated_data.pop('generator')
         # Secondary foreign table data
-        gen_object = self.create_handler_instance(gen_data)
-        tsd_object = self.create_handler_instance(tsd_data)
+        gen_object = self.create_handler(**gen_data)
+        tsd_object = self.create_handler(**tsd_data)
 
         # Create model instances
-        # gen_object = Handler.objects.create(**gen_data)
-        # tsd_object = Handler.objects.create(**tsd_data)
         manifest = Manifest.objects.create(generator=gen_object,
                                            tsd=tsd_object,
                                            **validated_data)
         for transporter in trans_data:
-            site_address_data = transporter.pop('site_address')
-            mail_address_data = transporter.pop('mail_address')
-            site_address = Address.objects.create(**site_address_data)
-            mail_address = Address.objects.create(**mail_address_data)
-            Transporter.objects.create(manifest=manifest, **transporter,
-                                       mail_address=mail_address,
-                                       site_address=site_address)
+            self.create_transporter(manifest, **transporter)
         for waste_line in waste_data:
             WasteLine.objects.create(manifest=manifest, **waste_line)
         return manifest
-
-    def create_handler_instance(self, handler_data: dict) -> Handler:
-        site_address_data = handler_data.pop('site_address')
-        mail_address_data = handler_data.pop('mail_address')
-        site_address = Address.objects.create(**site_address_data)
-        mail_address = Address.objects.create(**mail_address_data)
-        new_handler = Handler.objects.create(site_address=site_address,
-                                             mail_address=mail_address,
-                                             **handler_data)
-        return new_handler
 
     # https://www.django-rest-framework.org/api-guide/serializers/#overriding-serialization-and-deserialization-behavior
     def to_representation(self, instance) -> str:
         data = super(ManifestSerializer, self).to_representation(instance)
         data['import'] = instance.import_flag
-        # remove null fields when serializing manifest
-        for field in self.fields:
-            try:
-                if data[field] is None:
-                    data.pop(field)
-            except KeyError:
-                pass
         return data
 
     def to_internal_value(self, data):
