+++ apps/trak/models/__init__.py
@@ -1,4 +1,5 @@
-from .handlers import Address, Handler, Site
+from .address import Address
+from .handler import Handler, Site
 from .manifest import Manifest
 from .transporter import Transporter
 from .waste import WasteLine
