+++ apps/trak/templates/trak/manifest_details.html
@@ -66,23 +66,23 @@
                   <div class="col">
                     <p class="font-weight-bold">Mailing Address</p>
                     <p>
-                      {{ manifest.generator.mail_street_number }}
-                      {{ manifest.generator.mail_address1 }}
-                      {{ manifest.generator.mail_address2 }}
-                      {{ manifest.generator.mail_city }}
-                      {{ manifest.generator.mail_state }},
-                      {{ manifest.generator.mail_zip }}
+                      {{ manifest.generator.mail_address.street_number }}
+                      {{ manifest.generator.mail_address.address1 }}
+                      {{ manifest.generator.mail_address.address2 }}
+                      {{ manifest.generator.mail_address.city }}
+                      {{ manifest.generator.mail_address.state }},
+                      {{ manifest.generator.mail_address.zip }}
                     </p>
                   </div>
                   <div class="col">
                     <p class="font-weight-bold">Site Address</p>
                     <p>
-                      {{ manifest.generator.site_street_number }}
-                      {{ manifest.generator.site_address1 }}
-                      {{ manifest.generator.site_address2 }}
-                      {{ manifest.generator.site_city }}
-                      {{ manifest.generator.site_state }},
-                      {{ manifest.generator.site_zip }}
+                      {{ manifest.generator.site_address.street_number }}
+                      {{ manifest.generator.site_address.address1 }}
+                      {{ manifest.generator.site_address.address2 }}
+                      {{ manifest.generator.site_address.city }}
+                      {{ manifest.generator.site_address.state }},
+                      {{ manifest.generator.site_address.zip }}
                     </p>
                   </div>
                 </div>
@@ -142,23 +142,23 @@
                   <div class="col">
                     <p class="font-weight-bold">Mailing Address</p>
                     <p>
-                      {{ manifest.tsd.mail_street_number }}
-                      {{ manifest.tsd.mail_address1 }}
-                      {{ manifest.tsd.mail_address2 }}
-                      {{ manifest.tsd.mail_city }}
-                      {{ manifest.tsd.mail_state }},
-                      {{ manifest.tsd.mail_zip }}
+                      {{ manifest.tsd.mail_address.street_number }}
+                      {{ manifest.tsd.mail_address.address1 }}
+                      {{ manifest.tsd.mail_address.address2 }}
+                      {{ manifest.tsd.mail_address.city }}
+                      {{ manifest.tsd.mail_address.state }},
+                      {{ manifest.tsd.mail_address.zip }}
                     </p>
                   </div>
                   <div class="col">
                     <p class="font-weight-bold">Site Address</p>
                     <p>
-                      {{ manifest.tsd.site_street_number }}
-                      {{ manifest.tsd.site_address1 }}
-                      {{ manifest.tsd.site_address2 }}
-                      {{ manifest.tsd.site_city }}
-                      {{ manifest.tsd.site_state }},
-                      {{ manifest.tsd.site_zip }}
+                      {{ manifest.tsd.site_address.street_number }}
+                      {{ manifest.tsd.site_address.address1 }}
+                      {{ manifest.tsd.site_address.address2 }}
+                      {{ manifest.tsd.site_address.city }}
+                      {{ manifest.tsd.site_address.state }},
+                      {{ manifest.tsd.site_address.zip }}
                     </p>
                   </div>
                 </div>
