+++ runhaz.sh
@@ -55,14 +55,14 @@ dump_fixtures(){
     exec_cmd="$base_cmd dumpdata"
     fixture_dir="$base_dir/tests/fixtures"
     fixture_cmd=(
-    "-e contenttypes -e auth.permission > $fixture_dir/test_data.json"
+    "-e contenttypes -e auth.permission -e admin.logentry -e sessions.session > $fixture_dir/test_data.json"
     "trak.WasteLine --pks=1 > $fixture_dir/test_waste_line.json"
     )
     for i in "${fixture_cmd[@]}"
     do
         eval "$exec_cmd $i"
     done
-    echo "Data successfully dumped"
+    echo "Data dumped"
 }
 
 load_fixtures() {
