--- tests/fixtures/site_handler_manifest.json
@@ -1,216 +1,0 @@
-[
-  {
-    "model": "trak.handler",
-    "pk": 1,
-    "fields": {
-      "epa_id": "VATESTRAN03",
-      "name": "VA TEST GEN 2021",
-      "modified": null,
-      "registered": null,
-      "contact": {
-        "test": "test"
-      },
-      "emergency_phone": null,
-      "electronic_signatures_info": null,
-      "gis_primary": false,
-      "can_esign": null,
-      "limited_esign": null,
-      "registered_emanifest_user": false,
-      "site_street_number": "123",
-      "site_address1": "S CRYSTAL DRIVE",
-      "site_address2": null,
-      "site_city": "Arlington",
-      "site_state": "VA",
-      "site_country": "US",
-      "site_zip": "22002",
-      "mail_street_number": null,
-      "mail_address1": "123",
-      "mail_address2": "S CRYSTAL DRIVE",
-      "mail_city": "Arlington",
-      "mail_state": "VA",
-      "mail_country": "US",
-      "mail_zip": "22002"
-    }
-  },
-  {
-    "model": "trak.handler",
-    "pk": 2,
-    "fields": {
-      "epa_id": "VATEST000002",
-      "name": "TEST TRANSPORTER 2 OF VA",
-      "modified": null,
-      "registered": null,
-      "contact": {
-        "test": "test"
-      },
-      "emergency_phone": null,
-      "electronic_signatures_info": null,
-      "gis_primary": false,
-      "can_esign": null,
-      "limited_esign": null,
-      "registered_emanifest_user": false,
-      "site_street_number": "2777",
-      "site_address1": "VA TEST TRAN 2021 ROAD",
-      "site_address2": "SW",
-      "site_city": "Arlington",
-      "site_state": "VA",
-      "site_country": "US",
-      "site_zip": "22002",
-      "mail_street_number": "2777",
-      "mail_address1": "S CRYSTAL DRIVE",
-      "mail_address2": "S CRYSTAL DRIVE",
-      "mail_city": "Arlington",
-      "mail_state": "VA",
-      "mail_country": "US",
-      "mail_zip": "22002"
-    }
-  },
-  {
-    "model": "trak.handler",
-    "pk": 3,
-    "fields": {
-      "epa_id": "VATESTTSDF03",
-      "name": "VA TEST TSDF 2021",
-      "modified": null,
-      "registered": null,
-      "contact": {
-        "test": "test"
-      },
-      "emergency_phone": null,
-      "electronic_signatures_info": null,
-      "gis_primary": false,
-      "can_esign": null,
-      "limited_esign": null,
-      "registered_emanifest_user": false,
-      "site_street_number": "2777",
-      "site_address1": "S CRYSTAL DRIVE",
-      "site_address2": "SW",
-      "site_city": "Arlington",
-      "site_state": "VA",
-      "site_country": "US",
-      "site_zip": "22002",
-      "mail_street_number": "2777",
-      "mail_address1": "123",
-      "mail_address2": "S CRYSTAL DRIVE",
-      "mail_city": "Arlington",
-      "mail_state": "VA",
-      "mail_country": "US",
-      "mail_zip": "22002"
-    }
-  },
-  {
-    "model": "trak.handler",
-    "pk": 4,
-    "fields": {
-      "epa_id": "VATESTGEN001",
-      "name": "TEST TRANSPORTER 2 OF VA",
-      "modified": null,
-      "registered": null,
-      "contact": {
-        "test": "Test"
-      },
-      "emergency_phone": null,
-      "electronic_signatures_info": null,
-      "gis_primary": false,
-      "can_esign": null,
-      "limited_esign": null,
-      "registered_emanifest_user": false,
-      "site_street_number": "2777",
-      "site_address1": "S CRYSTAL DRIVE",
-      "site_address2": "SW",
-      "site_city": "Arlington",
-      "site_state": "VA",
-      "site_country": "US",
-      "site_zip": "22002",
-      "mail_street_number": "2777",
-      "mail_address1": "123",
-      "mail_address2": "S CRYSTAL DRIVE",
-      "mail_city": "Arlington",
-      "mail_state": "VA",
-      "mail_country": "US",
-      "mail_zip": "22002"
-    }
-  },
-  {
-    "model": "trak.handler",
-    "pk": 5,
-    "fields": {
-      "epa_id": "testid1",
-      "name": "a name",
-      "modified": null,
-      "registered": null,
-      "contact": {
-        "test": "test"
-      },
-      "emergency_phone": null,
-      "electronic_signatures_info": null,
-      "gis_primary": false,
-      "can_esign": null,
-      "limited_esign": null,
-      "registered_emanifest_user": false,
-      "site_street_number": "2777",
-      "site_address1": "S CRYSTAL DRIVE",
-      "site_address2": null,
-      "site_city": "Arlington",
-      "site_state": "VA",
-      "site_country": "US",
-      "site_zip": "22002",
-      "mail_street_number": "2777",
-      "mail_address1": "S CRYSTAL DRIVE",
-      "mail_address2": null,
-      "mail_city": "Arlington",
-      "mail_state": "VA",
-      "mail_country": "US",
-      "mail_zip": "22002"
-    }
-  },
-  {
-    "model": "trak.manifest",
-    "pk": 1,
-    "fields": {
-      "created_date": "2022-05-24T18:45:53.835Z",
-      "update_date": "2022-05-24T18:45:53.835Z",
-      "mtn": "000000001ELC",
-      "status": "InTransit",
-      "submission_type": "FullElectronic",
-      "signature_status": null,
-      "origin_type": "Service",
-      "shipped_date": "2022-05-24T18:45:09Z",
-      "potential_ship_date": "2022-05-24T18:45:14Z",
-      "received_date": null,
-      "certified_date": null,
-      "certified_by": null,
-      "generator": 1,
-      "tsd": 3,
-      "broker": null,
-      "rejection": false,
-      "rejection_info": null,
-      "discrepancy": false,
-      "residue": false,
-      "residue_new_mtn": [],
-      "import_flag": false,
-      "import_info": null,
-      "contains_residue_or_rejection": null,
-      "printed_document": null,
-      "form_document": null,
-      "additional_info": null,
-      "correction_info": null,
-      "ppc_status": null,
-      "locked": null,
-      "locked_reason": null,
-      "transfer_requested": null,
-      "transfer_status": null,
-      "original_sub_type": "FullElectronic",
-      "transfer_count": null,
-      "next_transfer_time": null
-    }
-  },
-  {
-    "model": "trak.site",
-    "pk": 1,
-    "fields": {
-      "name": "My Site",
-      "epa_site": 4
-    }
-  }
-]
